+++ src/renderer/index.tsx
@@ -39,6 +39,7 @@ import type {
   BrokerConfigRequest,
   ClusterHealthRequest,
   DeleteTopicRequest,
+  DeleteTopicsRequest,
   DiscoverRequest,
   GroupDetailRequest,
   GroupsRequest,
@@ -165,6 +166,8 @@ export default class KafkaExtensionRenderer extends Renderer.LensExtension {
     this.client.produce({ clusterId: this.clusterId, ...request });
   private readonly deleteTopic = (request: DeleteTopicRequest) =>
     this.client.deleteTopic({ clusterId: this.clusterId, ...request });
+  private readonly deleteTopics = (request: DeleteTopicsRequest) =>
+    this.client.deleteTopics({ clusterId: this.clusterId, ...request });
   private readonly resetOffsets = (request: ResetOffsetsRequest) =>
     this.client.resetOffsets({ clusterId: this.clusterId, ...request });
   private readonly schemaSubjectNames = (request: SchemaSubjectNamesRequest) => this.client.schemaSubjectNames(request);
@@ -262,6 +265,7 @@ export default class KafkaExtensionRenderer extends Renderer.LensExtension {
             messagesBrowse={this.messagesBrowse}
             produce={this.produce}
             deleteTopic={this.deleteTopic}
+            deleteTopics={this.deleteTopics}
             schemaRegistrySettings={this.schemaRegistrySettings}
             endpointSecrets={this.endpointSecrets}
             reachability={this.reachability}